+++ extension/src/externalresources/SharedAccessSignature.ts
